+++ src/routes/Router.tsx
@@ -1,5 +1,5 @@
 import { createBrowserRouter } from 'react-router-dom';
-import Main from '../pages/Main';
+import Main from '../pages/Main/Main';
 import Login from '../pages/Login/Login';
 import SignUp from '../pages/SignUp/SignUp';
 import MyPage from '../pages/MyPage';
